[Feat] #23 UIImageView extension을 삭제 후 setImage 메소드를 URLUIImageView에 구현
